Update tests to Python3 and use responses library to mock
Updated the tests to enable the mocking of responses to work correctly under Python3. Found an issue with the status code checking when mocking the session class of Requests.
Made use of Responses to mock the POST method of the requests library. The test coverage may need reviewing as it's not as well covered (specifically the keyword function which splits
up keywords into arguments for the Open311 protocol. Future work may include improved test coverage against the spec, as a way of ensuring compliance in the client library.